working on comments homework #2
