+++ assets/js/main.js
@@ -2,17 +2,18 @@ let count = 1;
 const counterEl = document.getElementById("counter");
 const resultEl = document.getElementById("result");
 
+
 function multiply(num) {
     return num * 129;
 }
 
-function increment() {
+export function increment() {
     count++;
     counterEl.textContent = count;
     resultEl.textContent = multiply(count);
 }
 
-function decrement() {
+export function decrement() {
     if (count > 0) {
         count--;
         counterEl.textContent = count;
