+++ index.html
@@ -7,8 +7,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="./assets/style/normalize.css">
     <link rel="stylesheet" href="./assets/style/style.css">
-    <script src="./assets/js/api.js"></script>
-    <script defer src="./assets/js/cardList.js"></script>
+    <script type="module" src="./assets/js/api.js"></script>
+    <script type="module" defer src="./assets/js/cardList.js"></script>
     <script defer src="./assets/js/search.js"></script>
 
     <title>catalog</title>
@@ -16,8 +16,6 @@
 
 <body>
     <header class="header">
-
-
         <form class="search-form ">
             <button type="submit" class="search-button"><img src="./assets/images/search.svg" alt="search"></button>
             <input type="text" class="search-input" placeholder="Search products">
