+++ product-details.html
@@ -7,8 +7,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="./assets/style/product-details.css">
     <link rel="stylesheet" href="./assets/style/style.css">
-    <script defer src="./assets/js/detail.js"></script>
-    <script defer src="./assets/js/api.js"></script>
+    <script type="module" defer src="./assets/js/detail.js"></script>
     <script defer src="./assets/js/search.js"></script>
     <title>detals</title>
 </head>
@@ -30,8 +29,7 @@
     </header>
     <div id="product-render" class="product-render">
     </div>
-    <script defer src="./assets/js/main.js">
-    </script>
+    <script type="module" src="./assets/js/main.js"></script>
 </body>
 
 </html>
